+++ ruter/ruter.go
@@ -1,4 +1,4 @@
-package sanntid
+package ruter
 
 import (
 	"fmt"
@@ -22,12 +22,11 @@ type sanntidArrivalData struct {
 	MonitoredVehicleJourney sanntidMonitoredVehicleJourney
 }
 
-func requestArrivalData(locationId int) ([]sanntidArrivalData, error) {
+func RequestArrivalData(locationId int) ([]sanntidArrivalData, error) {
 	var data []sanntidArrivalData
 
 	url := fmt.Sprintf("http://reisapi.ruter.no/stopvisit/getdepartures/%d", locationId)
 	res, err := goreq.Request{ Uri: url }.Do()
-
 	if err == nil {
 		res.Body.FromJsonTo(&data)
 	}
